requirements.txt
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
asyncua>=1.0.0
redis>=5.0.0
flask>=3.0.0
flask-socketio>=5.3.0
sqlalchemy>=2.0.0
matplotlib>=3.7.0
PyQt6>=6.5.0

